changes to calcula Corrente

+++ sensorCorrente.h
@@ -1,4 +1,3 @@
-
 #ifndef sensorCorrente_h
 #define sensorCorrente_h
 
@@ -8,22 +7,23 @@ enum sensorCorrente_type {T5B,T20A,T30A};
 
 
 class sensorCorrente{
-    
+
     public:
         sensorCorrente(int _pino,sensorCorrente_type modelo);
         float calculaRMS();
-        float calculaCorrente(int A);
-        void calibrar(); 
+        float calculaCorrente(float A);
+        void calibrar();
 	   float calculaDigitalIpp();
-        
+
+
 
 
     private:
         int pino;
-        int fatorConversao;
+        float fatorConversao;
         int zero = 512;
-    
-  
+
+
 };
